+++ dataquality-statistics/CHANGELOG.md
@@ -8,7 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 N/A
 ### Changed
-N/A
+- workitem([TPRUN-354](https://jira.talendforge.org/browse/TPRUN-354)): Improve "native" discovery for typed data
+
 ### Removed
 N/A
 ### Deprecated